libselinux: use size_t for index
The variable is used as array index and the only caller passes already
size_t.

Acked-by: Stephen Smalley <stephen.smalley.work@gmail.com>

+++ libselinux/src/booleans.c
@@ -329,9 +329,9 @@ int security_commit_booleans(void)
 		return -1;
 }
 
-static void rollback(SELboolean *boollist, int end)
+static void rollback(SELboolean *boollist, size_t end)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < end; i++)
 		security_set_boolean(
